requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
lxml==4.6.3
numpy==1.21.1
pandas==1.3.0
Pillow==8.3.1
praw==7.3.0
prawcore==2.2.0
pyaml==20.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.0
